refactor back to separate workers
